// web/src/api/generated/models/salesCheckResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */
import type { SalesChargeAdjustmentResponse } from './salesChargeAdjustmentResponse';
import type { SalesChargeAllocationResponse } from './salesChargeAllocationResponse';
import type { SalesPaymentResponse } from './salesPaymentResponse';
import type { SalesRefundResponse } from './salesRefundResponse';

export interface SalesCheckResponse {
  allocations?: SalesChargeAllocationResponse[];
  balance_vnd?: number;
  base_charge_vnd?: number;
  charge_adjustments?: SalesChargeAdjustmentResponse[];
  charge_vnd?: number;
  created_at?: string;
  effective_received_vnd?: number;
  id?: string;
  merged_into_check_id?: string;
  payments?: SalesPaymentResponse[];
  pending_refund_vnd?: number;
  refunds?: SalesRefundResponse[];
  state?: string;
  total_applied_vnd?: number;
  total_refunded_vnd?: number;
  total_voided_vnd?: number;
}